Redact against a specific pattern, also don't count occurrences
base## patterns introduced several patterns that cross-match heavily, especially
when test.check tries it. "22222..." is a pretty unlikely base64 string, but
test.check doesn't know that (and doesn't care or arguably even actively cares
that it's unlikely). But that's also a valid base16 and base32 string, so
dependent on pattern order in the default rules, it'd get redacted differently.
That's not really a bug in the behavior, but a bug in the test. This change
tests the same behavior but fixes redaction to a single rule so we know which
rule is applying.

Arguably, the pattern should not be part of the set of generated test results,
and we should actually `doseq` this. There is no meaningful way to minimize the
pattern. But since it's not broken, why fix it?

To show why count-occurrences is not a good test, consider this example that
test.check came up with:

 redacted "arn:aws:ar::89450:G",
 redacted-fixed-val "ar",
 orig-fixed-val "ar",