Remove base32-re-check
This test is fundamentally flawed for the same reason the duplicate group test
previously was: just because you're generating a sample from the base32-re
doesn't mean it'll unambiguously be base32; e.g. "222..." (a string test.check
loves to generate) is valid base16, base32, base64 and so inspect fails to
identify it but that doesn't _really_ mean failure.